chore: split file for NativeUtils
